feat: Add SQLBase and EngineSingleton classes
This commit adds the `SQLBase` and `EngineSingleton` classes to the `pzsm.db` module. The `SQLBase` class is a subclass of `DeclarativeBase` from `sqlalchemy.orm` and serves as the base class for all SQLAlchemy models in the project. The `EngineSingleton` class is responsible for creating and managing a singleton instance of the SQLAlchemy engine, using the `create_engine` function from `sqlalchemy.engine`.